feat: update the VIP number and add description
