--- vips/vip-357/bscmainnet.ts
@@ -1,63 +1,0 @@
-import { parseUnits } from "ethers/lib/utils";
-
-import { LzChainId, ProposalType } from "../../src/types";
-import { makeProposal } from "../../src/utils";
-
-const XVS_BRIDGE_ADMIN = "0x70d644877b7b73800E9073BCFCE981eAaB6Dbc21";
-
-export const MIN_DST_GAS = "300000";
-export const SINGLE_SEND_LIMIT = parseUnits("10000", 18);
-export const MAX_DAILY_SEND_LIMIT = parseUnits("50000", 18);
-export const SINGLE_RECEIVE_LIMIT = parseUnits("10200", 18);
-export const MAX_DAILY_RECEIVE_LIMIT = parseUnits("51000", 18);
-export const ZKSYNC_TRUSTED_REMOTE = "0x16a62b534e09a7534cd5847cfe5bf6a4b0c1b116";
-
-const vip357 = () => {
-  const meta = {
-    version: "v2",
-    title: "VIP-357 Enable BSC -> ZKSYNC bridge",
-    description: `#### Summary Enable BSC -> ZKSYNC bridge`,
-    forDescription: "I agree that Venus Protocol should proceed with this proposal",
-    againstDescription: "I do not think that Venus Protocol should proceed with this proposal",
-    abstainDescription: "I am indifferent to whether Venus Protocol proceeds or not",
-  };
-
-  return makeProposal(
-    [
-      {
-        target: XVS_BRIDGE_ADMIN,
-        signature: "setTrustedRemoteAddress(uint16,bytes)",
-        params: [LzChainId.zksyncmainnet, ZKSYNC_TRUSTED_REMOTE],
-      },
-      {
-        target: XVS_BRIDGE_ADMIN,
-        signature: "setMinDstGas(uint16,uint16,uint256)",
-        params: [LzChainId.zksyncmainnet, 0, MIN_DST_GAS],
-      },
-      {
-        target: XVS_BRIDGE_ADMIN,
-        signature: "setMaxDailyLimit(uint16,uint256)",
-        params: [LzChainId.zksyncmainnet, MAX_DAILY_SEND_LIMIT],
-      },
-      {
-        target: XVS_BRIDGE_ADMIN,
-        signature: "setMaxSingleTransactionLimit(uint16,uint256)",
-        params: [LzChainId.zksyncmainnet, SINGLE_SEND_LIMIT],
-      },
-      {
-        target: XVS_BRIDGE_ADMIN,
-        signature: "setMaxDailyReceiveLimit(uint16,uint256)",
-        params: [LzChainId.zksyncmainnet, MAX_DAILY_RECEIVE_LIMIT],
-      },
-      {
-        target: XVS_BRIDGE_ADMIN,
-        signature: "setMaxSingleReceiveTransactionLimit(uint16,uint256)",
-        params: [LzChainId.zksyncmainnet, SINGLE_RECEIVE_LIMIT],
-      },
-    ],
-    meta,
-    ProposalType.REGULAR,
-  );
-};
-
-export default vip357;
